+++ pytsite/widget/__init__.py
@@ -37,7 +37,8 @@ def _init():
     assetman.js_module('pytsite-widget-select-color-picker', __name__ + '@js/color-picker')
     assetman.js_module('pytsite-widget-misc-bootstrap-table', __name__ + '@js/bootstrap-table')
 
-    assetman.preload(__name__ + '@js/init-widgets.js', True)
+    assetman.preload(__name__ + '@css/widget.css', permanent=True)
+    assetman.preload(__name__ + '@js/init-widgets.js', permanent=True)
 
 
 _init()